HDFS-727. bug setting block size hdfsOpenFile. Contributed by Eli Collins.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hdfs/trunk@882708 13f79535-47bb-0310-9956-ffa450edef68

+++ src/c++/libhdfs/hdfs.c
@@ -679,7 +679,7 @@ hdfsFile hdfsOpenFile(hdfsFS fs, const char* path, int flags,
         if (!blockSize) {
             if (invokeMethod(env, &jVal, &jExc, INSTANCE, jConfiguration, 
                              HADOOP_CONF, "getLong", "(Ljava/lang/String;J)J",
-                             jStrBlockSize, 67108864)) {
+                             jStrBlockSize, (jlong)67108864)) {
                 errno = errnoFromException(jExc, env, "org.apache.hadoop.conf."
                                            "FileSystem::%s(%s)", method,
                                            signature);
